FEA VotingRegressor (#12513)
